Add animation for navbar
