+++ api/lib/src/models/tool.dart
@@ -1,6 +1,3 @@
-import 'dart:math';
-
-import '../converter/core.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 
 part 'tool.freezed.dart';
@@ -16,19 +13,4 @@ class ToolOption with _$ToolOption {
 
   factory ToolOption.fromJson(Map<String, dynamic> json) =>
       _$ToolOptionFromJson(json);
-}
-
-@freezed
-class ToolState with _$ToolState {
-  const factory ToolState({
-    @Default(false) bool rulerEnabled,
-    @Default(false) bool gridEnabled,
-    @DoublePointJsonConverter()
-    @Default(Point(0.0, 0.0))
-    Point<double> rulerPosition,
-    @Default(0) double rulerAngle,
-  }) = _ToolState;
-
-  factory ToolState.fromJson(Map<String, dynamic> json) =>
-      _$ToolStateFromJson(json);
 }
