+++ api/pubspec.yaml
@@ -17,3 +17,4 @@ dev_dependencies:
   freezed: ^2.3.4
   lints: ^2.1.0
   test: ^1.24.3
+  analyzer: 5.12.0
