+++ app/lib/cubits/current_index.dart
@@ -19,6 +19,7 @@ import 'package:xml/xml.dart';
 
 import '../embed/embedding.dart';
 import '../handlers/handler.dart';
+import '../models/tool.dart';
 import '../models/viewport.dart';
 import '../selections/selection.dart';
 import '../theme.dart';
