+++ app/lib/renderers/renderer.dart
@@ -23,6 +23,7 @@ import '../cubits/current_index.dart';
 import '../cubits/transform.dart';
 import '../helpers/xml_helper.dart';
 import '../models/label.dart';
+import '../models/tool.dart';
 
 part 'backgrounds/box.dart';
 part 'backgrounds/empty.dart';
