+++ app/lib/view_painter.dart
@@ -5,6 +5,7 @@ import 'package:butterfly_api/butterfly_api.dart';
 import 'package:flutter/material.dart';
 
 import 'cubits/transform.dart';
+import 'models/tool.dart';
 import 'selections/selection.dart';
 
 class ForegroundPainter extends CustomPainter {
